# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user's orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of user orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order from cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrderDetails
      tags: [orders]
      summary: Get single order details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (Admin Simulator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdateInput"
      responses:
        "200":
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /pandits:
    get:
      operationId: getPandits
      tags: [pandits]
      summary: Get all pandits
      responses:
        "200":
          description: List of all pandits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pandit"
    post:
      operationId: createPandit
      tags: [pandits]
      summary: Add a new pandit (Admin CRUD)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanditInput"
      responses:
        "201":
          description: Pandit created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pandit"
...
  /pandits/{id}:
    put:
      operationId: updatePandit
      tags: [pandits]
      summary: Update pandit details (Admin CRUD)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanditInput"
      responses:
        "200":
          description: Pandit updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pandit"
    delete:
      operationId: deletePandit
      tags: [pandits]
      summary: Delete pandit (Admin CRUD)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pandit deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItem:
      type: object
      properties:
        name:
          type: string
        qty:
          type: integer
        price:
          type: integer
        unit:
          type: string
      required:
        - name
        - qty
        - price
        - unit
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        orderId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        amount:
          type: integer
        delivery:
          type: integer
        status:
          type: string
        addressText:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - orderId
        - items
        - amount
        - delivery
        - status
        - addressText
        - createdAt
        - updatedAt
    OrderInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        amount:
          type: integer
        delivery:
          type: integer
        addressText:
          type: string
      required:
        - items
        - amount
        - delivery
        - addressText
    StatusUpdateInput:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Pandit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        specialty:
          type: string
        category:
          type: string
        rating:
          type: number
        experience:
          type: string
        bookings:
          type: integer
        age:
          type: integer
        city:
          type: string
        address:
          type: string
        available:
          type: string
        specializations:
          type: array
          items:
            type: string
        muhurats:
          type: array
          items:
            type: string
        poojas:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              duration:
                type: string
              price:
                type: integer
              includesPrasad:
                type: boolean
            required:
              - id
              - name
              - duration
              - price
              - includesPrasad
        initials:
          type: string
        avatarColor:
          type: string
        email:
          type: string
        password:
          type: string
        imageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - shortName
        - specialty
        - category
        - rating
        - experience
        - bookings
        - age
        - city
        - address
        - available
        - specializations
        - muhurats
        - poojas
        - initials
        - avatarColor
        - email
        - password
        - createdAt
        - updatedAt
    PanditInput:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        specialty:
          type: string
        category:
          type: string
        rating:
          type: number
        experience:
          type: string
        bookings:
          type: integer
        age:
          type: integer
        city:
          type: string
        address:
          type: string
        available:
          type: string
        specializations:
          type: array
          items:
            type: string
        muhurats:
          type: array
          items:
            type: string
        poojas:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              duration:
                type: string
              price:
                type: integer
              includesPrasad:
                type: boolean
            required:
              - id
              - name
              - duration
              - price
              - includesPrasad
        initials:
          type: string
        avatarColor:
          type: string
        email:
          type: string
        password:
          type: string
        imageUrl:
          type: string
      required:
        - name
        - shortName
        - specialty
        - category
        - rating
        - experience
        - bookings
        - age
        - city
        - address
        - available
        - specializations
        - muhurats
        - poojas
        - initials
        - avatarColor
        - email
        - password