# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  userId: number;
  orderId: string;
  items: OrderItem[];
  amount: number;
  delivery: number;
  status: string;
  addressText: string;
  createdAt: Date;
  updatedAt: Date;
}