# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface OrderInput {
  items: OrderItem[];
  amount: number;
  delivery: number;
  addressText: string;
}