# lib/api-zod/src/generated/types/pandit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PanditPoojasItem } from './panditPoojasItem';

export interface Pandit {
  id: number;
  name: string;
  shortName: string;
  specialty: string;
  category: string;
  rating: number;
  experience: string;
  bookings: number;
  age: number;
  city: string;
  address: string;
  available: string;
  specializations: string[];
  muhurats: string[];
  poojas: PanditPoojasItem[];
  initials: string;
  avatarColor: string;
  email: string;
  password: string;
  imageUrl?: string;
  createdAt: Date;
  updatedAt: Date;
}